Add hybrid dashboard combining data analytics and AI insights
- Created HybridDashboardView with comprehensive analytics
- Added AI insights carousel with detailed views
- Implemented metric cards with trends and sparklines
- Added interactive charts (nutrition distribution, weekly trends)
- Included AI recommendations with actionable items
- Created detailed nutrient analytics table
- Integrated Spoonacular API for food recognition
- Updated DashboardSelectorView to include new hybrid option
- Added comprehensive documentation for Spoonacular setup

The hybrid dashboard combines the best features from the data-focused
dashboard (option 3) and AI insights dashboard (option 4), providing
users with both detailed analytics and personalized AI recommendations.

+++ ios/HealthTracker/Views/AddMenuView.swift
@@ -376,7 +376,7 @@ struct ExerciseSearchView: View {
         newExercise.id = UUID()
         newExercise.name = exercise.name
         newExercise.type = exercise.type.rawValue
-        newExercise.duration = duration
+        newExercise.duration = Int32(duration)
         newExercise.caloriesBurned = exercise.caloriesPerMinute * duration
         newExercise.date = selectedDate
         newExercise.timestamp = Date()
